Change Main, Command to work as intended
Replace Command's ArrayList to a Map<String, Command> to make Command
lookup more efficient.

Change Main main method to exit on "exit", processCommand on receiving
a List of Strings whose first element is not a blank String ("").
Uses a while(true) loop which may not be the best course of action,
but removes the seemingly infinite loop of while(scanner.hasNext()).
This was due to the scanner blocking and asking System.in for another
input, hence always returned true.

Add InvalidCommandException class - a checked exception for when the
List of Strings does not match the String names of any of the Commands.

Todo - make InvalidCommandException give more meaningful error messages
       (and possibly recover by asking for a new input)

Todo - implement subclasses of Command for specific commands, adding
       their functionality