Move functionality from Main to Command
Main had very similar structure to Command which warranted the change.

Change InvalidCommandException to provide a value for the last Command.
This requires that the InvalidCommandException takes the last Command
as a parameter.

Change main method in Command class to include a try catch statement.
This ensures the user is warned about entering an incorrect command,
and lists the valid Commands from the last Commands given.