Implement Add and Cancel Commands
Change Command class following issues in implementing AddCommand and
CancelCommand.
Add two new fields, a List of Commands, previousCommands and a boolean,
keepLooping.
The boolean keepLooping is used insted of true in the while loop of the
getInput method. This allows the loop to be broken and control to
return to the previous Command.
The List previousCommands is used so that each prior Command could be
accessed. This is of critical importance in CancelCommand where each
previous command (except the base command) needs to have keepLooping
set to false.
Change the parameters that use previousCommand to use previousCommands.
Add method getInput to clean up main method.

Add AddCommand. AddCommand serves as the superclass for all other
Commands wishing to add something to the Company. It's execute method
prints out valid subcommands as add itself is not a final Command.
This functionality may be moved into an overwritten help method when
the HelpCommand is implemented.

Add CancelCommand. CancelCommand is a final Command that breaks out of
wherever the user is and returns them to the base Command (Command that
runs when the program starts running). For example a user may be at the
AddCommand but want to return to the beginning of the program. A
StepBack Command could also be implemented to walk back just one
Command instead of right to the beginning of the program.

Todo - move main method into separate control class.
Todo - add HelpCommand and attach to all Commands.
Todo - provide helpful information after CancelCommand has been called.
Todo - add StepBackCommand to return to the previously issued Command.