Add Control class and split files into packages
Add the Control class that contains static methods getInput and
processCommand for controlling the program control flow.
Remove from Control methods the List of Commands parameter
previousCommands. Details about the route taken to the current Command
are located in the Command object itself. The current Command is given
as a parameter instead of previousCommands.

Refactor code into packages and make compatible with change to Control
methods.

Closes #1