Add HelpCommand
The HelpCommand is given as a Command to most other Commands (really
all but exit) and calls that Command's help function. This now means
that anything the user types will be checked against the current
Command's subCommands (instead of evaluating the String directly). This
is neater and will allow other Commands to be added more easily.

Add resetPreviousCommands method in class Command. HelpCommand uses
this method to clear its List of previousCommands so that when it is
called, the path will always be analysed.

Change Command and its children to use HelpCommand where appropriate.
Format the information given to the user better so the program is
easier to use.

Closes #2
Closes #3