Add StepBackCommand class
The StepBackCommand allows the user to return to the previously issued
Command without resetting the program completely.

Change AddCommand to include a StepBackCommand (even though the same
functionality is supplied by CancelCommand).

Add description of command and company variables to javadoc of
Control's processCommand method.

Fix bug in processCommand method of class Control. The case that the
method is to be called again recursively (because the user has entered
more than one Command) needs the newCommand's subcommands, not the
current command's subcommands.

Closes #4