Add AddPersonCommand
The AddPersonCommand adds a Person to the Company. The user is prompted
for the Person's name, address and Share information (numberOfShares
and dividend). The Share information is converted from a String to an
int and double respectively. Exceptions are used to catch obvious
NumberFormatExceptions that could result.

Add constructor to AddCommand class

A constructor that takes a string is required so AddPersonCommand can
extend by giving its name.

Todo - Add a DocumentFilter to allow only certain characters into the
       name and address
Todo - Check for valid (but incorrect) input (i.e. if the dividend per
       share is too large, or too many shares etc.)
Todo - Automatic capitalisation of name input